2DRenderer: Mark all regions of deleted node as dirty (Backport from Andy Nichols <andy.nichols@qt.io>)
Previous behavior was to always subtract the current bounding rect of
items when calculating the previous dirty region.  When you remove a
node though, there is no current bounding rect because the item no
longer exists.  The last valid boundingRect was use instead, which led
to the previous bounding rect to not get marked as dirty when the node
was removed.