Software Image Node: Do not make equality check in setTexture (Backport from Andy Nichols <andy.nichols@qt.io>)
When setTexture was called on the Software image node, we were doing a
check to see if the QSGTexture pointer was the same as the current
texture assuming it meant that if they were the same pointer then it was
in fact the same texture, and thus not dirty.  This was a wrong
assumption, because the texture data can be different but still use the
same pointer.  So the node should always respond by updating the texture
when setTexture is called and mark the nodes material as dirty for the
renderer.

This allows canvas items to be updated at runtime with the software
renderer.