Software Adaptation: Improve clip node handling & Software Adaptation: Fix issue with clip node handling (Backport from Andy Nichols <andy.nichols@qt.io>)
This commit solves two issues that were coming up with nested clip
nodes.  The first was that determining the intersection of two or more
nested clip nodes was calculated incorrectly.  This was fixed by making
sure the calculated clip region was in the correct coordinate space.

The second issue was because there was no distinction between an empty
unclipped region, and a completely clipped region.  So a property was
added to specify when a RenderableNode had a clipRegion so that the
empty QRegion could be interpreted correctly.  This makes sure that we
don't end up drawing nodes that were culled.

Software Adaptation: Fix issue with clip node handling

Under certain circumstances the clip node state would not get
propagated properly to child nodes.