Fix crash when using Software Renderer on HiDPI QBackingStore (Backport from Andy Nichols <andy.nichols@qt.io>)
For some reason the QPaintDevice of a QBackingStore is not valid on
HiDPI QBackingStores until QBackingStore::begin() has been called.  I
suspect this is a bug in qtbase, but until this is fixed this change
will make sure that the paint device will always be valid, even in this
unusual situation.